Risolto errore visualizzazione finestra swing
