--- src/main/java/com/mycompany/swing/ImgFrame.java
@@ -1,134 +1,0 @@
-/*
- * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
- * Click nbfs://nbhost/SystemFileSystem/Templates/GUIForms/JFrame.java to edit this template
- */
-package com.mycompany.swing;
-
-import com.mycompany.exception.DimException;
-import com.mycompany.exception.ImgException;
-import java.awt.BorderLayout;
-import java.util.logging.Level;
-import java.util.logging.Logger;
-import javax.swing.ImageIcon;
-
-/**
- * Questa classe ImgFrame prende in input un path che deve corrisppndere ad un'immagine
- * ed un eventuale stringa che sarà possibile visualizzare non appena ci si poggia sulla foto con il cursore,
- * verrà creata una finestra swing che visualizzerà quell'immagine;
- * nel momento in cui il path specificato non è un'img viene lanciata una ImgException, la dimensione della foto
- * deve rientrare in certi parametri di pixel, visualizzare le costanti di questa classe per saperne di più.
- * @author santo
- */
-public class ImgFrame extends javax.swing.JFrame {
-    
-    public static final Integer WIDTH_MIN = 600;  // larghezza
-    public static final Integer HEIGHT_MIN = 413; // altezza
-    public static final Integer WIDTH_MAX = 1200;
-    public static final Integer HEIGHT_MAX = 825;
-    private javax.swing.JLabel jLabel1;
-    
-    
-    public ImgFrame(String pathImg) throws ImgException, DimException {       
-        this(pathImg, "");
-    }
-    
-    
-    /**
-     * 
-     * @param pathImg
-     * @param LabelToolTipText, questa stringa verrà visualizzata non appena ci si posizione con il cursore sulla Label
-     * @throws ImgException 
-     * @throws acazzo.DimException 
-     */
-    public ImgFrame(String pathImg, String LabelToolTipText) throws ImgException, DimException {              
-        if(CheckImg.isImage(pathImg)) {
-            ImageIcon imgIcon = new ImageIcon(pathImg);           
-            if(imgIcon.getIconWidth() < WIDTH_MIN || imgIcon.getIconWidth() > WIDTH_MAX || imgIcon.getIconHeight() < HEIGHT_MIN || imgIcon.getIconHeight() > HEIGHT_MAX) {
-                throw new DimException();
-            } else {   
-                initComponents();
-                myInit(imgIcon, LabelToolTipText);
-            }                
-        } else {
-            throw new ImgException();
-        }    
-    }
-    
-
-    private void myInit(ImageIcon imgIcon, String LabelToolTipText) {
-        jLabel1 = new javax.swing.JLabel(imgIcon); 
-        jLabel1.setSize(100, 50);
-        jLabel1.setText("");
-        if(!LabelToolTipText.isEmpty())
-            jLabel1.setToolTipText(LabelToolTipText);
-        getContentPane().add(jLabel1, BorderLayout.CENTER);  // agiungo al frame la Label
-    }
-   
-    
-    /**
-     * This method is called from within the constructor to initialize the form.
-     * WARNING: Do NOT modify this code. The content of this method is always
-     * regenerated by the Form Editor.
-     */
-    @SuppressWarnings("unchecked")
-    // <editor-fold defaultstate="collapsed" desc="Generated Code">//GEN-BEGIN:initComponents
-    private void initComponents() {
-
-        setDefaultCloseOperation(javax.swing.WindowConstants.EXIT_ON_CLOSE);
-        setTitle("Img_Joe_Ellie");
-        setSize(new java.awt.Dimension(0, 0));
-
-        pack();
-    }// </editor-fold>//GEN-END:initComponents
-
-    /**
-     * @param args the command line arguments
-     */
-    public static void main(String args[]) throws DimException {
-
-        try {
-            /* Set the Nimbus look and feel */
-            //<editor-fold defaultstate="collapsed" desc=" Look and feel setting code (optional) ">
-            /* If Nimbus (introduced in Java SE 6) is not available, stay with the default look and feel.
-            * For details see http://download.oracle.com/javase/tutorial/uiswing/lookandfeel/plaf.html
-            */
-            try {
-                for (javax.swing.UIManager.LookAndFeelInfo info : javax.swing.UIManager.getInstalledLookAndFeels()) {
-                    if ("Nimbus".equals(info.getName())) {
-                        javax.swing.UIManager.setLookAndFeel(info.getClassName());
-                        break;
-                    }
-                }
-            } catch (ClassNotFoundException ex) {
-                java.util.logging.Logger.getLogger(ImgFrame.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
-            } catch (InstantiationException ex) {
-                java.util.logging.Logger.getLogger(ImgFrame.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
-            } catch (IllegalAccessException ex) {
-                java.util.logging.Logger.getLogger(ImgFrame.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
-            } catch (javax.swing.UnsupportedLookAndFeelException ex) {
-                java.util.logging.Logger.getLogger(ImgFrame.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
-            }
-            //</editor-fold>
-            //</editor-fold>
-            //</editor-fold>
-            //</editor-fold>
-            
-            /* Create and display the form */
-            /*java.awt.EventQueue.invokeLater(new Runnable() {
-            public void run() {
-            new ImgFrame().setVisible(true);
-            }
-            });*/    
-            ImgFrame ig = new ImgFrame("resources//img//fotoSoggiorno960x660.jpg", "quando si stava bene...");
-            ig.setVisible(true);    // quando viene eseguita questa istruzione, appare la finestra                 
-        } catch (ImgException ex) {
-            Logger.getLogger(ImgFrame.class.getName()).log(Level.SEVERE, null, ex);
-        }
-        //</editor-fold> 
-        //</editor-fold> 
-        //</editor-fold> 
-        //</editor-fold> 
-    }
-    // Variables declaration - do not modify//GEN-BEGIN:variables
-    // End of variables declaration//GEN-END:variables
-}
